[7G.3] Split output/console.py into console/ sub-package
phi_scan/output/console.py (1307 LOC) split into five focused modules:
- core.py: console instance, _detect_unicode_support, progress, spinner
- findings.py: findings table, file tree, code context, category breakdown
- summary.py: banner, phase separators, scan header, clean/violation panels
- baseline.py: baseline summary, diff, drift warning, scan notice panels
- __init__.py: constants, _UNICODE_SUPPORTED, _resolve_symbol; re-exports
  all public symbols with `name as name` for backwards compatibility

_UNICODE_SUPPORTED and _resolve_symbol remain in __init__.py to preserve
monkeypatch compatibility in existing tests. pyproject.toml gains an I001
per-file-ignore for __init__.py to suppress isort splitting re-export blocks.
