+++ CONTRIBUTING.md
@@ -67,7 +67,12 @@ make help       Show all targets
 | `phi_scan/fhir_recognizer.py` | Layer 3 — FHIR R4 field-name scanning |
 | `phi_scan/hl7_scanner.py` | Layer 3 — HL7 v2 segment parsing (optional) |
 | `phi_scan/detection_coordinator.py` | Combines layers, deduplicates findings |
-| `phi_scan/output/console.py` | Rich terminal UI — `display_*`, banner, progress |
+| `phi_scan/output/console/` | Rich terminal UI — `display_*`, banner, progress (sub-package) |
+| `phi_scan/output/console/__init__.py` | Constants, `_UNICODE_SUPPORTED`, `_resolve_symbol`; re-exports all public symbols |
+| `phi_scan/output/console/core.py` | Shared infrastructure: `_rich_console`, `get_console()`, `create_scan_progress`, `display_status_spinner` |
+| `phi_scan/output/console/findings.py` | Findings table, file tree, code context panel, category breakdown |
+| `phi_scan/output/console/summary.py` | Banner, phase separators, scan header, clean/violation summary panels |
+| `phi_scan/output/console/baseline.py` | Baseline summary, diff, drift warning, scan notice panels |
 | `phi_scan/output/serializers.py` | Pure-data serialisers — JSON, CSV, SARIF, JUnit, GitLab |
 | `phi_scan/output/dashboard.py` | Live dashboard layout builders |
 | `phi_scan/output/watch.py` | File-watcher event UI |
@@ -83,6 +88,7 @@ make help       Show all targets
 ### Where to Add Things
 
 - **New output format** (JSON, SARIF, etc.): `phi_scan/output/serializers.py` → export from `phi_scan/output/__init__.py` → add to `OutputFormat` in `constants.py` → wire in `cli.py`
+- **New terminal display function**: add to the appropriate `phi_scan/output/console/` sub-module (`findings.py`, `summary.py`, or `baseline.py`) → re-export from `phi_scan/output/console/__init__.py` using `name as name` syntax
 - **New CI/CD platform**: `phi_scan/ci_integration.py` — add a `_HttpRequestConfig`-based call using `_execute_http_request`; wire into `post_pr_comment` and `set_commit_status` dispatch tables
 - **New PHI pattern**: `phi_scan/regex_detector.py` — add pattern + validator; update `PhiCategory` in `constants.py` if a new HIPAA category
 - **New CLI command**: `phi_scan/cli.py` — follow the existing Typer app/sub-app pattern
